# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenEMIS Core",
        "description": "The [OpenEMIS](https://www.openemis.org/) initiative aims to deploy a high-quality Education Management Information System (EMIS) designed to collect and report data on schools, students, teachers and staff. The system was conceived by `UNESCO` to be a royalty-free system that can be easily customized to meet the specific needs of member countries.",
        "termsOfService": "https://www.openemis.org/terms-of-service/",
        "contact": {
            "email": "support@openemis.org"
        },
        "license": {
            "name": "GNU General Public License V3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://demo.openemis.org/core"
        }
    ],
    "paths": {
        "/api/v4/exams/{examId}": {
            "get": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Get details of a specific exam",
                "description": "Returns details of the exam identified by the provided examId",
                "operationId": "a601e38c583fced7c8c6f9cbb92e8872",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "description": "ID of the exam",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "PSLE"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Primary School Leaving Examination"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "registration_start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2020-01-01"
                                                },
                                                "registration_end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2020-07-31"
                                                },
                                                "academic_period_id": {
                                                    "type": "integer",
                                                    "example": "29"
                                                },
                                                "education_grade_id": {
                                                    "type": "integer",
                                                    "example": "64"
                                                },
                                                "modified_user_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "modified": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "created_user_id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "created": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2020-03-06 05:10:34"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unsuccessful."
                    }
                }
            }
        },
        "/api/v4/exams/{examId}/centres/{centreId}": {
            "get": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Get centre details for a specific exam",
                "description": "Returns details of the centre associated with the specified exam",
                "operationId": "d5e05c7613a869402879ca0992cb9287",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "description": "ID of the exam",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "centreId",
                        "in": "path",
                        "description": "ID of the centre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Avory Primary School"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "P1002"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "270 Duke Lane"
                                                },
                                                "postal_code": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "contact_person": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "83948723"
                                                },
                                                "fax": {
                                                    "type": "string",
                                                    "example": "83948723"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "contact@avoryprimary.com"
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "example": "avoryprimary.com"
                                                },
                                                "institution_id": {
                                                    "type": "integer",
                                                    "example": "6"
                                                },
                                                "area_id": {
                                                    "type": "integer",
                                                    "example": "11"
                                                },
                                                "academic_period_id": {
                                                    "type": "integer",
                                                    "example": "29"
                                                },
                                                "examination_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "total_registered": {
                                                    "type": "integer",
                                                    "example": "0"
                                                },
                                                "modified_user_id": {
                                                    "type": "integer",
                                                    "example": null
                                                },
                                                "modified": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": null
                                                },
                                                "created_user_id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "created": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2020-03-06 05:13:01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unsuccessful."
                    }
                }
            }
        },
...
    }
}